feat(gateway): 重构网关过滤器实现完整的请求验证流程
- 替换 getAppInfoHosts 方法为 getAppHostByAppId，通过 appId 查询应用主机
- 新增完整的请求参数验证包括 accessKey、nonce、timestamp 和签名验证
- 实现基于 appId 的接口路由转发功能
- 添加详细的请求日志记录和错误响应处理
- 统一错误响应格式并优化异常处理机制
- 修改 InterfaceInfo 实体中 appId 类型从 Integer 到 Long
- 更新相关 DTO 和服务实现适配新的字段类型
- 移除 Dubbo 依赖相关的 Maven 配置
- 优化数据库查询性能，添加查询条件限制

+++ MoLiAPI/moli-api-backend/src/main/java/com/bingli/MoliAPI/model/dto/interfaceInfo/InterfaceInfoAddRequest.java
@@ -13,8 +13,7 @@ public class InterfaceInfoAddRequest implements Serializable {
 /**
  * 应用 id
  */
-    private Integer appId;
-
+    private Long appId;
 
     /**
      * 名称
